SWF Molder Network Function

Attemps tomodifies a habitat graph to increase the cover of a specified category

swf_molderN(
  g,
  swfCat,
  agriCat,
  iterations,
  Q,
  NNeighbors = 1,
  swfCover = 0.3,
  max_radius = 1,
  np = 1
)

Arguments

g

Graph object representing the habitat.

swfCat

Category value in the graph to be increased.

agriCat

Category value in the graph to allocate.

iterations

Number of iterations to run the modification process.

Q

Number of nodes to be potentially modified in each iteration.

NNeighbors

Number of agriCat neighbor nodes to consider for allocation.

swfCover

Desired cover proportion for the swfCat category.

max_radius

Maximum search radius for neighbours.

np

Number of cores for parallel processing.